refactor: Day를 유틸리티 클래스로 변경하고, 불필요한 객체 생성을 없앰.
